% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allmethods.R
\name{allmethods}
\alias{allmethods}
\title{Test All Methods}
\usage{
allmethods(
  x = ggplot2::diamonds$price,
  y = ggplot2::diamonds$depth,
  big.title = FALSE
)
}
\arguments{
\item{x}{List or column of numeric values.}

\item{y}{List or column of numeric values.}

\item{big.title}{Set equal to TRUE to have a bigger title.}
}
\value{
A ggplot2 plot.
}
\description{
Test All Methods
}
\examples{
diamonds <- ggplot2::diamonds
allmethods(x = diamonds$carat, y = diamonds$depth)

}
