% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierplot.R
\name{outlierplot}
\alias{outlierplot}
\title{Create Outlier plot}
\usage{
outlierplot(
  x,
  y,
  labels = NULL,
  alpha = NULL,
  withcolor = TRUE,
  notheme = FALSE,
  detection.method = "PCS",
  default.title = FALSE
)
}
\arguments{
\item{x}{List or column of numeric values.}

\item{y}{List or column of numeric values.}

\item{labels}{List or column of string labels.}

\item{alpha}{Transparency of scatterplot points.}

\item{withcolor}{Set to FALSE for black and white graph. Defalts to TRUE.}

\item{notheme}{Set to false to avoid setting theme.}

\item{detection.method}{Method of outlier detection ("HDo", "PCS", adjOut", "DDC", "MCD")}

\item{default.title}{Set to true to have title be name of outlier detection method and number of outliers found.}
}
\value{
A ggplot2 plot.
}
\description{
Create Outlier plot
}
\examples{
diamonds <- ggplot2::diamonds
outlierplot(x = diamonds$carat, y = diamonds$depth)
outlierplot(x = diamonds$carat, y = diamonds$depth, labels = diamonds$cut)
outlierplot(x = diamonds$carat, y = diamonds$depth, alpha = .5, withcolor = FALSE)

}
