% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierplotold.R
\name{outlierplotold}
\alias{outlierplotold}
\title{Create Outlier plot}
\usage{
outlierplotold(x, y, labels = NULL, alpha = NULL)
}
\arguments{
\item{x}{List or column of numeric values.}

\item{y}{List or column of numeric values.}

\item{labels}{List or column of string labels.}

\item{alpha}{Transparency of scatterplot points.}
}
\value{
A ggplot2 plot.
}
\description{
Create Outlier plot
}
\examples{
diamonds <- ggplot2::diamonds
outlierplotold(x = diamonds$carat, y = diamonds$depth)
outlierplotold(x = diamonds$carat, y = diamonds$depth, labels = diamonds$cut)

}
