% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierplotold2.R
\name{outlierplotold2}
\alias{outlierplotold2}
\title{Create Outlier plot}
\usage{
outlierplotold2(x, y, labels = NULL, alpha = NULL, withcolor = TRUE)
}
\arguments{
\item{x}{List or column of numeric values.}

\item{y}{List or column of numeric values.}

\item{labels}{List or column of string labels.}

\item{alpha}{Transparency of scatterplot points.}

\item{withcolor}{Set to FALSE for black and white graph. Defalts to TRUE.}
}
\value{
A ggplot2 plot.
}
\description{
Create Outlier plot
}
\examples{
diamonds <- ggplot2::diamonds
outlierplot(x = diamonds$carat, y = diamonds$depth)
outlierplot(x = diamonds$carat, y = diamonds$depth, labels = diamonds$cut)
outlierplot(x = diamonds$carat, y = diamonds$depth, alpha = .5, withcolor = FALSE)

}
